Fix construction of VCI RPC management errors
Closes: VRVDR-49736

For non-Go VCI clients (C++, Python), errors passed back over the DBUS by the likes of check() calls are not being properly displayed, and end up looking something like this:

```
[policy ingress-map]

Request could not be completed because the requested operation failed for some reason not covered by any other error condition.

[[policy ingress-map]] failed.
```

The problem is in the construction of the mgmtError object - the 'info' field is assigned to with the error text embedded, leading it to not be extracted on the receiving end.  The fix is to assign the information direct to the Message field, in the same way that Go components do.  A failed validation now looks like this:

```
[policy ingress-map]

incomplete ingress-map

[[policy ingress-map]] failed.
```

I've verified that a C++ and Python client that previously failed now work.